+++ src/room007/exploration/stats.py
@@ -4,12 +4,12 @@
 from collections import Counter
 from functools import partial
 from itertools import chain
-from os.path import basename, join, splitext
-import re
+from os.path import basename, join
 
 import pandas as pd
 
 from room007.data import info
+from room007.exploration.common import *
 
 
 TAG_STATS_IDX = pd.Index(('present', 'total'))
@@ -28,21 +28,6 @@ def tags_stat(frame, stat=num_words):
     return stat(frame['tags'])
 
 
-def norm_tag(tag):
-    return tag.replace('-', ' ')
-
-
-def tag_in_text(tag, text):
-    tag_re = '\\b{tag}\\b'.format(tag=re.escape(tag))
-    return re.search(tag_re, text) is not None
-
-
-def question_mentions_tag(qrec, tag):
-    normed = norm_tag(tag)
-    return (tag_in_text(normed, qrec['title']) or
-            tag_in_text(normed, qrec['content']))
-
-
 def rec_tag_stats(rec):
     num_tags_in_text = sum(map(partial(question_mentions_tag, rec),
                                rec['tags']))
@@ -53,8 +38,8 @@ def rec_tag_stats(rec):
 
 
 def frame_tag_stats(frame):
-    # TODO Do this splitting earlier.
     return frame.apply(rec_tag_stats, 1).sum()
+
 
 def tag_available(frame):
     frame['scontent'] = frame['content'].str.split()
@@ -72,8 +57,8 @@ if __name__ == "__main__":
                               index=TAG_STATS_IDX,
                               dtype='i4')
     overall_stats2 = pd.Series((0, 0),
-                              index=TAG_STATS_IDX,
-                              dtype='i4')
+                               index=TAG_STATS_IDX,
+                               dtype='i4')
     data_info = info.CleanedData()
     dataframes = info.get_train_dataframes(data_info)
     for fname, data in dataframes.items():
